+++ service.py
@@ -19,7 +19,8 @@ app = Flask(__name__)
 CORS(app, resources={
     r"/guardar": {"origins": "http://localhost:4200"},
     r"/reconocer": {"origins": "http://localhost:4200"},
-    r"/recorteFacial": {"origins": "http://localhost:4200"}
+    r"/recorteFacial": {"origins": "http://localhost:4200"},
+    r"/borrar_contenido": {"origins": "http://localhost:4200"}
 })
 
 # Conectarse a la base de datos de MongoDB
@@ -54,6 +55,44 @@ def create_object():
         return jsonify({"message": "Objeto guardado correctamente", "object_id": object_id}), 201
     else:
         return jsonify({"message": "Error al guardar el objeto"}), 500
+    
+def borrar_contenido_carpeta(carpeta):
+    # Verificar si la carpeta existe
+    if os.path.exists(carpeta):
+        # Obtener lista de archivos en la carpeta
+        archivos = os.listdir(carpeta)
+
+        # Borrar cada archivo en la carpeta
+        for archivo in archivos:
+            ruta_archivo = os.path.join(carpeta, archivo)
+            os.remove(ruta_archivo)
+    else:
+        print(f"La carpeta {carpeta} no existe")
+
+@app.route('/borrar_contenido', methods=['GET'])
+def borrar_contenido():
+    try:
+        # Rutas de las dos carpetas que quieres vaciar
+        carpeta1 = "./faces"
+        carpeta2 = "./archivosjson"
+        carpeta3 = "./imagen"
+
+        # Borrar contenido de las dos carpetas
+        borrar_contenido_carpeta(carpeta1)
+        borrar_contenido_carpeta(carpeta2)
+        borrar_contenido_carpeta(carpeta3)
+
+        # Devolver respuesta de éxito
+        resp = jsonify({'server': "Clean Server"})
+        resp.status_code = 201
+        return resp
+    
+    except Exception as e:
+        # Devolver respuesta en caso de error
+        resp = jsonify({'server': "Error al limpiar el servidor", 'message': str(e)})
+        resp.status_code = 201
+        return resp
+
 
 @app.route('/reconocer', methods=['POST', 'OPTIONS'])
 def reconocer():
@@ -108,13 +147,15 @@ def reconocer():
         response.status_code = 500
         return response
 
+
 @app.route('/recorteFacial', methods=['POST'])
 def upload_file():
     try:
-        # Lista para almacenar los objetos JSON
+
+        # Lista para almacenar los objetos JSON---
         json_list = []
 
-        # check if the post request has the file part
+       # check if the post request has the file part
         if 'file' not in request.files:
             resp = jsonify({'message' : 'No file part in the request'})
             resp.status_code = 400
@@ -129,9 +170,9 @@ def upload_file():
         #obtenemos el nombre del archivo
         filename = secure_filename(imagesPath.filename)
         #asignamos una ruta al archivo
-        output_file = "./imagen/" + filename
+        ruta_imagen  = "./imagen/" + filename
         #guardamos el archivo subido
-        imagesPath.save(output_file)
+        imagesPath.save(ruta_imagen )
 
         if not os.path.exists("faces"):
             os.makedirs("faces")
@@ -152,79 +193,77 @@ def upload_file():
         )
 
         # Cargar imagen
-        image = cv2.imread(output_file)
+        image = cv2.imread(ruta_imagen)
+
+        if image is None:
+            print(f"Error al leer la imagen en la ruta: {ruta_imagen}")
+            resp = jsonify({'error': 'Error al leer la imagen en la ruta'})
+            resp.status_code = 500
+            return resp
 
         # Convertir imagen de BGR a RGB
-        #print(image.shape)
+        print(image.shape)
         image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
-        plt.imshow(image)
-        #plt.show()
 
         # Detectar caras en la imagen
         boxes, _ = mtcnn.detect(image)
-    
 
-        try:
-            # Aquí se realiza el corte de las imágenes
-            if boxes is not None:
-                i = 1
-                for box in boxes:
-                    box = [int(coord) for coord in box]
-                    face_image_encodings = face_recognition.face_encodings(image, known_face_locations=[box])[0]
-                    #print(face_image_encodings)
-                    face = image[box[1]:box[3], box[0]:box[2], :]
-                    cv2.imwrite("faces/" + str(i) + ".png", cv2.cvtColor(face, cv2.COLOR_BGR2RGB))
-                    #i += 1
-                    #cv2.imshow("face", cv2.cvtColor(face, cv2.COLOR_BGR2RGB))
-                    cv2.waitKey(0)
-                    
-                    
-                    #------------Codificar Code: Codificamos cada una de las imagenes------------
-                    output_file_faces = "./faces/" + str(i) + ".png"
-                    imagen_file = cv2.imread(output_file_faces)
+        # Aquí se realiza el corte de las imágenes
+        if boxes is not None:
+            i = 1
+            for box in boxes:
+                box = [int(coord) for coord in box]
+                face = image[box[1]:box[3], box[0]:box[2], :]
+                cv2.imwrite("faces/" + str(i) + ".png", cv2.cvtColor(face, cv2.COLOR_BGR2RGB))
+                i += 1
+            
+            a = 1
+            for contador in boxes:
+                output_file_faces = "./faces/" + str(a) + ".png"
+                print("output_file_faces: ", output_file_faces)
+                imagen_file = cv2.imread(output_file_faces)
 
-                    face_locations = face_recognition.face_locations(imagen_file, number_of_times_to_upsample=2, model="hog")
-                    print("FaceLocation: ", len(face_locations))
-                    
-                    imgBase64 = base64.b64encode(open('./faces/' + str(i) + ".png", 'rb').read()).decode('utf-8')
-                    i += 1
-                    if len(face_locations) >= 0:
-                        face_loc = face_locations[0]
-                    
-                        face_image_encodings = face_recognition.face_encodings(imagen_file, known_face_locations=[face_loc])[0]
-                        #print("face_image_encondings:", face_image_encodings)
+                face_locations = face_recognition.face_locations(imagen_file,  number_of_times_to_upsample=2, model="hog")
+                print("FaceLocation: ", len(face_locations))
 
 
-                        if not os.path.exists("archivosjson"):
-                            os.makedirs("archivos")
+                imgBase64 = base64.b64encode(open(output_file_faces, 'rb').read()).decode('utf-8')
+                if len(face_locations) > 0:
+                    face_loc = face_locations[0]
+                    face_image_encodings = face_recognition.face_encodings(imagen_file, known_face_locations=[face_loc])[0]
+                
 
-                        # Obtener el nombre base del archivo sin la extensión
-                        base_name = os.path.splitext(os.path.basename(output_file_faces))[0]
-                        print("base_name:", base_name)
-                        
-                        # Crear diccionario con la información de la cara
-                        face_data = {
-                            "encoding": face_image_encodings.tolist(),
-                            'imagen': imgBase64,
-                        }
-                        #print("face_dict:", face_data)
+                    # Crear diccionario con la información de la cara
+                    face_data = {
+                        "encodings": face_image_encodings.tolist(),
+                        'imagen': imgBase64,  
+                    }
+                    #print("face_dict:", face_data)
 
-                        cv2.rectangle(imagen_file, (face_loc[3], face_loc[0]),(face_loc[1], face_loc[2]), (0, 255, 0))
-                        
-                        # Guardar diccionario en archivo json en el directorio "archivos"
-                        with open(f"archivosjson/{base_name}.json", "w") as f:
-                            json.dump(face_data, f)
-                            
-                        # Agregar el JSON a la lista
+                    if not os.path.exists("archivosjson"):
+                        os.makedirs("archivos")
+
+                    # Obtener el nombre base del archivo sin la extensión
+                    # base_name = os.path.splitext(os.path.basename(image))[0]
+                    base_name = str(a)
+
+                    # Guardar diccionario en archivo json en el directorio "archivos"
+                    with open(f"archivosjson/{base_name}.json", "w") as f:
+                        json.dump(face_data, f)
+
+                    # Agregar el JSON a la lista
                         json_list.append(face_data)
-                        
-                    else:
-                        resp = jsonify({'error': "No se ha detectado ninguna cara en la imagen"})
-                        resp.status_code = 201
-                        return resp
-        except Exception as e:
-            print("Error-Bucle:", str(e))
-        
+
+                else:
+                    print("No se ha detectado ninguna cara en la imagen")
+                    # imgBase64 = base64.b64encode(open(output_file_faces, 'rb').read()).decode('utf-8')
+                    face_data = {
+                        "encodings": [0],
+                        'imagen': imgBase64,  
+                    }
+                    json_list.append(face_data)
+                a += 1
+
         resp = jsonify({'server': json_list})
         resp.status_code = 201
         return resp
